requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.37
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
